backend/requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-cors-headers==3.3.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-sslify==0.2.7
djangorestframework==3.11.0
httpie==2.1.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mod-wsgi==4.7.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==7.1.2
pycodestyle==2.6.0
Pygments==2.6.1
pylint==2.5.2
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
